requirements.txt
fastapi
uvicorn
databases
sqlalchemy
python-dotenv